fix: address review finding #007 — extract block business logic to service
Move haversine filtering, resolution rate computation, average days
calculation, and top-issue aggregation from server/routes/block.ts
to server/services/block.ts, consistent with other domain services.
